requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.11.1
black==24.2.0
bleach==5.0.1
blinker==1.6.2
bokeh==3.2.0
build==1.2.1
cachetools==4.1.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
DateTime==5.1
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
Django==5.0.3
dnspython==2.6.1
email_validator==2.1.1
entrypoints==0.4
environs==11.0.0
et-xmlfile==1.1.0
executing==0.9.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.16.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==23.5.26
flutter==0.1
fonttools==4.34.4
frozendict==2.4.0
frozenlist==1.3.3
fsspec==2024.5.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.6.6
google-api-core==1.34.1
google-api-python-client==1.9.1
google-auth==2.32.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-speech==2.26.1
google-generativeai==0.7.2
googleapis-common-protos==1.63.0
graphviz==0.20.3
grpcio==1.64.0
grpcio-status==1.48.2
gspread==3.6.0
h11==0.14.0
hellosign-python-sdk==4.0.0
html5lib==1.1
httpcore==0.16.3
httplib2==0.18.1
httptools==0.6.1
httpx==0.23.3
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywebrtc==0.6.0
ipywidgets==8.0.7
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.9.0
jupyter==1.0.0
jupyter-bokeh==3.0.7
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
kubernetes==29.0.0
Levenshtein==0.25.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.21.2
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
mmh3==4.1.0
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.0
notebook==6.4.12
notion-client==2.2.1
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
onnx==1.14.0
openai==1.30.5
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.6.0.66
openpyxl==3.0.10
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.0.0
pandas-stubs==1.5.2.221213
pandas_ta==0.3.14b0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.3
peewee==3.17.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2023.11.15
platformdirs==4.1.0
plotly==5.22.0
plotly-express==0.4.1
posthog==3.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
pycocotools==2.0.6
pycodestyle==2.9.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pynndescent==0.5.12
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.1.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
python-pygaze==0.7.6
python-telegram-bot==20.0
pytz==2023.3
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
rapidfuzz==3.8.1
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==13.7.0
rsa==4.0
sagemath==1.3.0
schedule==1.1.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.11.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
Send2Trash==1.8.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.14.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.3.0
starlette==0.37.2
statsmodels==0.13.5
streamlit==1.31.1
sympy==1.11.1
tenacity==8.2.3
terminado==0.15.0
threadpoolctl==3.5.0
tiktoken==0.3.1
tinycss2==1.1.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.15.2
tornado==6.2
tqdm==4.66.4
traitlets==5.3.0
typer==0.12.3
types-pytz==2022.7.0.0
typing_extensions==4.12.0
tzdata==2023.3
tzlocal==5.2
ujson==5.10.0
umap-learn==0.5.6
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.29.0
uvloop==0.19.0
validators==0.22.0
virtualenv==20.25.0
watchfiles==0.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
whisper==1.1.10
whitenoise==6.6.0
widgetsnbextension==4.0.8
wrapt==1.16.0
xyzservices==2023.5.0
yarl==1.8.2
yfinance==0.2.36
zipp==3.17.0
zope.interface==6.0
